--- a/benchmark/profile.js
+++ b/benchmark/profile.js
@@ -14,6 +14,8 @@ var md = require('../')({
 // var data = fs.readFileSync(path.join(__dirname, '/samples/lorem1.txt'), 'utf8');
 var data = fs.readFileSync(path.join(__dirname, '../test/fixtures/commonmark/spec.txt'), 'utf8');
 
-for (var i = 0; i < 20; i++) {
+console.time('profile');
+for (var i = 0; i < 200; i++) {
   md.render(data);
 }
+console.timeEnd('profile');