--- a/lib/common/utils.js
+++ b/lib/common/utils.js
@@ -13,6 +13,14 @@ function has(object, key) {
   return _hasOwnProperty.call(object, key);
 }
 
+var _repeat = String.prototype.repeat || function (count) {
+  return new Array(count + 1).join (this);
+};
+
+function repeat(str, count) {
+  return _repeat.call (str, count);
+}
+
 // Merge objects
 //
 function assign(obj /*from1, from2, from3, ...*/) {
@@ -237,7 +245,7 @@ function isMdAsciiPunct(ch) {
   }
 }
 
-// Hepler to unify [reference labels].
+// Helper to unify [reference labels].
 //
 function normalizeReference(str) {
   // Trim and collapse whitespace
@@ -302,6 +310,7 @@ exports.lib.ucmicro         = require('uc.micro');
 exports.assign              = assign;
 exports.isString            = isString;
 exports.has                 = has;
+exports.repeat              = repeat;
 exports.unescapeMd          = unescapeMd;
 exports.unescapeAll         = unescapeAll;
 exports.isValidEntityCode   = isValidEntityCode;