--- a/lib/common/utils.js
+++ b/lib/common/utils.js
@@ -128,25 +128,38 @@ function unescapeAll(str) {
 
 ////////////////////////////////////////////////////////////////////////////////
 
-var HTML_ESCAPE_TEST_RE = /[&<>"]/;
-var HTML_ESCAPE_REPLACE_RE = /[&<>"]/g;
-var HTML_REPLACEMENTS = {
-  '&': '&amp;',
-  '<': '&lt;',
-  '>': '&gt;',
-  '"': '&quot;'
-};
+var escapeHtml = (function () {
+  if ( typeof XMLSerializer === 'function' ) {
+    var serializer = new XMLSerializer ();
+    var attr = document.createAttribute ( 'attr' );
+    var re = /[&<>"]/;
+    return function escapeHtml ( str ) {
+      if ( !re.test ( str ) ) return str;
+      attr.value = str;
+      return serializer.serializeToString ( attr );
+    }
+  } else {
+    var HTML_ESCAPE_TEST_RE = /[&<>"]/;
+    var HTML_ESCAPE_REPLACE_RE = /[&<>"]/g;
+    var HTML_REPLACEMENTS = {
+      '&': '&amp;',
+      '<': '&lt;',
+      '>': '&gt;',
+      '"': '&quot;'
+    };
 
-function replaceUnsafeChar(ch) {
-  return HTML_REPLACEMENTS[ch];
-}
+    function replaceUnsafeChar(ch) {
+      return HTML_REPLACEMENTS[ch];
+    }
 
-function escapeHtml(str) {
-  if (HTML_ESCAPE_TEST_RE.test(str)) {
-    return str.replace(HTML_ESCAPE_REPLACE_RE, replaceUnsafeChar);
+    return function escapeHtml(str) {
+      if (HTML_ESCAPE_TEST_RE.test(str)) {
+        return str.replace(HTML_ESCAPE_REPLACE_RE, replaceUnsafeChar);
+      }
+      return str;
+    }
   }
-  return str;
-}
+})();
 
 ////////////////////////////////////////////////////////////////////////////////