--- a/lib/rules_block/code.js
+++ b/lib/rules_block/code.js
@@ -4,9 +4,9 @@
 
 
 module.exports = function code(state, startLine, endLine/*, silent*/) {
-  var nextLine, last, token;
-
   if (state.sCount[startLine] - state.blkIndent < 4) { return false; }
+
+  var nextLine, last, token;
 
   last = nextLine = startLine + 1;