--- a/lib/rules_block/fence.js
+++ b/lib/rules_block/fence.js
@@ -4,13 +4,13 @@
 
 
 module.exports = function fence(state, startLine, endLine, silent) {
+  // if it's indented more than 3 spaces, it should be a code block
+  if (state.sCount[startLine] - state.blkIndent >= 4) { return false; }
+
   var marker, len, params, nextLine, mem, token, markup,
       haveEndMarker = false,
       pos = state.bMarks[startLine] + state.tShift[startLine],
       max = state.eMarks[startLine];
-
-  // if it's indented more than 3 spaces, it should be a code block
-  if (state.sCount[startLine] - state.blkIndent >= 4) { return false; }
 
   if (pos + 3 > max) { return false; }