--- a/lib/rules_block/hr.js
+++ b/lib/rules_block/hr.js
@@ -3,15 +3,16 @@
 'use strict';
 
 var isSpace = require('../common/utils').isSpace;
+var repeat = require('../common/utils').repeat;
 
 
 module.exports = function hr(state, startLine, endLine, silent) {
+  // if it's indented more than 3 spaces, it should be a code block
+  if (state.sCount[startLine] - state.blkIndent >= 4) { return false; }
+
   var marker, cnt, ch, token,
       pos = state.bMarks[startLine] + state.tShift[startLine],
       max = state.eMarks[startLine];
-
-  // if it's indented more than 3 spaces, it should be a code block
-  if (state.sCount[startLine] - state.blkIndent >= 4) { return false; }
 
   marker = state.src.charCodeAt(pos++);
 
@@ -39,7 +40,7 @@ module.exports = function hr(state, startLine, endLine, silent) {
 
   token        = state.push('hr', 'hr', 0);
   token.map    = [ startLine, state.line ];
-  token.markup = Array(cnt + 1).join(String.fromCharCode(marker));
+  token.markup = repeat (String.fromCharCode(marker), cnt);
 
   return true;
 };