--- a/lib/rules_block/html_block.js
+++ b/lib/rules_block/html_block.js
@@ -21,14 +21,14 @@ var HTML_SEQUENCES = [
 
 
 module.exports = function html_block(state, startLine, endLine, silent) {
-  var i, nextLine, token, lineText,
-      pos = state.bMarks[startLine] + state.tShift[startLine],
-      max = state.eMarks[startLine];
+  if (!state.md.options.html) { return false; }
 
   // if it's indented more than 3 spaces, it should be a code block
   if (state.sCount[startLine] - state.blkIndent >= 4) { return false; }
 
-  if (!state.md.options.html) { return false; }
+  var i, nextLine, token, lineText,
+      pos = state.bMarks[startLine] + state.tShift[startLine],
+      max = state.eMarks[startLine];
 
   if (state.src.charCodeAt(pos) !== 0x3C/* < */) { return false; }