--- a/lib/rules_block/list.js
+++ b/lib/rules_block/list.js
@@ -11,7 +11,6 @@ function skipBulletListMarker(state, startLine) {
   var marker, pos, max, ch;
 
   pos = state.bMarks[startLine] + state.tShift[startLine];
-  max = state.eMarks[startLine];
 
   marker = state.src.charCodeAt(pos++);
   // Check bullet
@@ -20,6 +19,8 @@ function skipBulletListMarker(state, startLine) {
       marker !== 0x2B/* + */) {
     return -1;
   }
+
+  max = state.eMarks[startLine];
 
   if (pos < max) {
     ch = state.src.charCodeAt(pos);
@@ -84,13 +85,14 @@ function skipOrderedListMarker(state, startLine) {
 }
 
 function markTightParagraphs(state, idx) {
-  var i, l,
+  var i, l, token,
       level = state.level + 2;
 
   for (i = idx + 2, l = state.tokens.length - 2; i < l; i++) {
-    if (state.tokens[i].level === level && state.tokens[i].type === 'paragraph_open') {
+    token = state.tokens[i];
+    if (token.level === level && token.type === 'paragraph_open') {
       state.tokens[i + 2].hidden = true;
-      state.tokens[i].hidden = true;
+      token.hidden = true;
       i += 2;
     }
   }
@@ -98,6 +100,21 @@ function markTightParagraphs(state, idx) {
 
 
 module.exports = function list(state, startLine, endLine, silent) {
+  // if it's indented more than 3 spaces, it should be a code block
+  if (state.sCount[startLine] - state.blkIndent >= 4) { return false; }
+
+  // Special case:
+  //  - item 1
+  //   - item 2
+  //    - item 3
+  //     - item 4
+  //      - this one is a paragraph continuation
+  if (state.listIndent >= 0 &&
+    state.sCount[startLine] - state.listIndent >= 4 &&
+    state.sCount[startLine] < state.blkIndent) {
+    return false;
+  }
+
   var ch,
       contentStart,
       i,
@@ -128,21 +145,6 @@ module.exports = function list(state, startLine, endLine, silent) {
       token,
       isTerminatingParagraph = false,
       tight = true;
-
-  // if it's indented more than 3 spaces, it should be a code block
-  if (state.sCount[startLine] - state.blkIndent >= 4) { return false; }
-
-  // Special case:
-  //  - item 1
-  //   - item 2
-  //    - item 3
-  //     - item 4
-  //      - this one is a paragraph continuation
-  if (state.listIndent >= 0 &&
-      state.sCount[startLine] - state.listIndent >= 4 &&
-      state.sCount[startLine] < state.blkIndent) {
-    return false;
-  }
 
   // limit conditions when list can interrupt
   // a paragraph (validation mode only)