--- a/lib/rules_block/reference.js
+++ b/lib/rules_block/reference.js
@@ -6,6 +6,9 @@ var isSpace              = require('../common/utils').isSpace;
 
 
 module.exports = function reference(state, startLine, _endLine, silent) {
+  // if it's indented more than 3 spaces, it should be a code block
+  if (state.sCount[startLine] - state.blkIndent >= 4) { return false; }
+
   var ch,
       destEndPos,
       destEndLineNo,
@@ -26,9 +29,6 @@ module.exports = function reference(state, startLine, _endLine, silent) {
       pos = state.bMarks[startLine] + state.tShift[startLine],
       max = state.eMarks[startLine],
       nextLine = startLine + 1;
-
-  // if it's indented more than 3 spaces, it should be a code block
-  if (state.sCount[startLine] - state.blkIndent >= 4) { return false; }
 
   if (state.src.charCodeAt(pos) !== 0x5B/* [ */) { return false; }