--- a/lib/rules_block/state_block.js
+++ b/lib/rules_block/state_block.js
@@ -4,6 +4,7 @@
 
 var Token = require('../token');
 var isSpace = require('../common/utils').isSpace;
+var repeat = require('../common/utils').repeat;
 
 
 function StateBlock(src, md, env, tokens) {
@@ -173,14 +174,13 @@ StateBlock.prototype.skipCharsBack = function skipCharsBack(pos, code, min) {
 
 // cut lines range from source.
 StateBlock.prototype.getLines = function getLines(begin, end, indent, keepLastLF) {
-  var i, lineIndent, ch, first, last, queue, lineStart,
+  var i, lineIndent, ch, first, last, lineStart,
+      lines = '',
       line = begin;
 
   if (begin >= end) {
-    return '';
+    return lines;
   }
-
-  queue = new Array(end - begin);
 
   for (i = 0; line < end; line++, i++) {
     lineIndent = 0;
@@ -215,13 +215,13 @@ StateBlock.prototype.getLines = function getLines(begin, end, indent, keepLastLF
     if (lineIndent > indent) {
       // partially expanding tabs in code blocks, e.g '\t\tfoobar'
       // with indent=2 becomes '  \tfoobar'
-      queue[i] = new Array(lineIndent - indent + 1).join(' ') + this.src.slice(first, last);
+      lines += repeat (' ', lineIndent - indent) + this.src.slice(first, last);
     } else {
-      queue[i] = this.src.slice(first, last);
+      lines += this.src.slice(first, last);
     }
   }
 
-  return queue.join('');
+  return lines;
 };
 
 // re-export Token class to use in block rules