--- a/lib/rules_block/table.js
+++ b/lib/rules_block/table.js
@@ -50,19 +50,19 @@ function escapedSplit(str) {
 
 
 module.exports = function table(state, startLine, endLine, silent) {
-  var ch, lineText, pos, i, l, nextLine, columns, columnCount, token,
-      aligns, t, tableLines, tbodyLines, oldParentType, terminate,
-      terminatorRules;
-
   // should have at least two lines
   if (startLine + 2 > endLine) { return false; }
 
-  nextLine = startLine + 1;
+  var nextLine = startLine + 1;
 
   if (state.sCount[nextLine] < state.blkIndent) { return false; }
 
   // if it's indented more than 3 spaces, it should be a code block
   if (state.sCount[nextLine] - state.blkIndent >= 4) { return false; }
+
+  var ch, lineText, pos, i, l, columns, columnCount, token,
+      aligns, t, tableLines, tbodyLines, oldParentType, terminate,
+      terminatorRules;
 
   // first character of the second line should be '|', '-', ':',
   // and no other characters are allowed but spaces;