--- a/lib/rules_core/linkify.js
+++ b/lib/rules_core/linkify.js
@@ -17,12 +17,12 @@ function isLinkClose(str) {
 
 
 module.exports = function linkify(state) {
+  if (!state.md.options.linkify) { return; }
+
   var i, j, l, tokens, token, currentToken, nodes, ln, text, pos, lastPos,
       level, htmlLinkLevel, url, fullUrl, urlText,
       blockTokens = state.tokens,
       links;
-
-  if (!state.md.options.linkify) { return; }
 
   for (j = 0, l = blockTokens.length; j < l; j++) {
     if (blockTokens[j].type !== 'inline' ||