--- a/lib/rules_core/normalize.js
+++ b/lib/rules_core/normalize.js
@@ -4,18 +4,21 @@
 
 
 // https://spec.commonmark.org/0.29/#line-ending
-var NEWLINES_RE  = /\r\n?|\n/g;
-var NULL_RE      = /\0/g;
+var CRLF_RE = /\r\n?/g;
 
 
 module.exports = function normalize(state) {
-  var str;
+  var src = state.src;
 
-  // Normalize newlines
-  str = state.src.replace(NEWLINES_RE, '\n');
+  // Normalize CRLF newlines
+  src = src.replace(CRLF_RE, '\n');
 
   // Replace NULL characters
-  str = str.replace(NULL_RE, '\uFFFD');
+  for (var i = 0, l = src.length; i < l; i++) {
+    if (!src.charCodeAt(i)) {
+      src = src.slice (0, i) + '\uFFFD' + src.slice (i + 1);
+    }
+  }
 
-  state.src = str;
+  state.src = src;
 };