--- a/lib/rules_core/smartquotes.js
+++ b/lib/rules_core/smartquotes.js
@@ -26,7 +26,7 @@ function process_inlines(tokens, state) {
   for (i = 0; i < tokens.length; i++) {
     token = tokens[i];
 
-    thisLevel = tokens[i].level;
+    thisLevel = token.level;
 
     for (j = stack.length - 1; j >= 0; j--) {
       if (stack[j].level <= thisLevel) { break; }