--- a/lib/rules_inline/autolink.js
+++ b/lib/rules_inline/autolink.js
@@ -9,10 +9,10 @@ var AUTOLINK_RE = /^([a-zA-Z][a-zA-Z0-9+.\-]{1,31}):([^<>\x00-\x20]*)$/;
 
 
 module.exports = function autolink(state, silent) {
+  if (state.src.charCodeAt(state.pos) !== 0x3C/* < */) { return false; }
+
   var url, fullUrl, token, ch, start, max,
       pos = state.pos;
-
-  if (state.src.charCodeAt(pos) !== 0x3C/* < */) { return false; }
 
   start = state.pos;
   max = state.posMax;