--- a/lib/rules_inline/backticks.js
+++ b/lib/rules_inline/backticks.js
@@ -4,11 +4,12 @@
 
 
 module.exports = function backtick(state, silent) {
-  var start, max, marker, token, matchStart, matchEnd, openerLength, closerLength,
-      pos = state.pos,
-      ch = state.src.charCodeAt(pos);
+  var ch = state.src.charCodeAt(state.pos);
 
   if (ch !== 0x60/* ` */) { return false; }
+
+  var start, max, marker, token, matchStart, matchEnd, openerLength, closerLength,
+      pos = state.pos;
 
   start = pos;
   pos++;