--- a/lib/rules_inline/emphasis.js
+++ b/lib/rules_inline/emphasis.js
@@ -6,13 +6,13 @@
 // Insert each marker as a separate text token, and add it to delimiter list
 //
 module.exports.tokenize = function emphasis(state, silent) {
-  var i, scanned, token,
-      start = state.pos,
-      marker = state.src.charCodeAt(start);
-
   if (silent) { return false; }
 
+  var marker = state.src.charCodeAt(state.pos);
+
   if (marker !== 0x5F /* _ */ && marker !== 0x2A /* * */) { return false; }
+
+  var i, scanned, token;
 
   scanned = state.scanDelims(state.pos, marker === 0x2A);