--- a/lib/rules_inline/entity.js
+++ b/lib/rules_inline/entity.js
@@ -13,9 +13,9 @@ var NAMED_RE   = /^&([a-z][a-z0-9]{1,31});/i;
 
 
 module.exports = function entity(state, silent) {
-  var ch, code, match, pos = state.pos, max = state.posMax;
+  if (state.src.charCodeAt(state.pos) !== 0x26/* & */) { return false; }
 
-  if (state.src.charCodeAt(pos) !== 0x26/* & */) { return false; }
+  var ch, code, match, pos = state.pos, max = state.posMax;
 
   if (pos + 1 < max) {
     ch = state.src.charCodeAt(pos + 1);