--- a/lib/rules_inline/escape.js
+++ b/lib/rules_inline/escape.js
@@ -13,9 +13,9 @@ for (var i = 0; i < 256; i++) { ESCAPED.push(0); }
 
 
 module.exports = function escape(state, silent) {
-  var ch, pos = state.pos, max = state.posMax;
+  if (state.src.charCodeAt(state.pos) !== 0x5C/* \ */) { return false; }
 
-  if (state.src.charCodeAt(pos) !== 0x5C/* \ */) { return false; }
+  var ch, pos = state.pos, max = state.posMax;
 
   pos++;