--- a/lib/rules_inline/html_inline.js
+++ b/lib/rules_inline/html_inline.js
@@ -14,17 +14,17 @@ function isLetter(ch) {
 
 
 module.exports = function html_inline(state, silent) {
-  var ch, match, max, token,
-      pos = state.pos;
-
   if (!state.md.options.html) { return false; }
 
   // Check start
-  max = state.posMax;
-  if (state.src.charCodeAt(pos) !== 0x3C/* < */ ||
-      pos + 2 >= max) {
+  var max = state.posMax;
+  if (state.src.charCodeAt(state.pos) !== 0x3C/* < */ ||
+      state.pos + 2 >= max) {
     return false;
   }
+
+  var ch, match, token,
+      pos = state.pos;
 
   // Quick fail on second char
   ch = state.src.charCodeAt(pos + 1);