--- a/lib/rules_inline/image.js
+++ b/lib/rules_inline/image.js
@@ -7,6 +7,9 @@ var isSpace              = require('../common/utils').isSpace;
 
 
 module.exports = function image(state, silent) {
+  if (state.src.charCodeAt(state.pos) !== 0x21/* ! */) { return false; }
+  if (state.src.charCodeAt(state.pos + 1) !== 0x5B/* [ */) { return false; }
+
   var attrs,
       code,
       content,
@@ -23,9 +26,6 @@ module.exports = function image(state, silent) {
       href = '',
       oldPos = state.pos,
       max = state.posMax;
-
-  if (state.src.charCodeAt(state.pos) !== 0x21/* ! */) { return false; }
-  if (state.src.charCodeAt(state.pos + 1) !== 0x5B/* [ */) { return false; }
 
   labelStart = state.pos + 2;
   labelEnd = state.md.helpers.parseLinkLabel(state, state.pos + 1, false);