--- a/lib/rules_inline/link.js
+++ b/lib/rules_inline/link.js
@@ -7,6 +7,8 @@ var isSpace              = require('../common/utils').isSpace;
 
 
 module.exports = function link(state, silent) {
+  if (state.src.charCodeAt(state.pos) !== 0x5B/* [ */) { return false; }
+
   var attrs,
       code,
       label,
@@ -22,8 +24,6 @@ module.exports = function link(state, silent) {
       max = state.posMax,
       start = state.pos,
       parseReference = true;
-
-  if (state.src.charCodeAt(state.pos) !== 0x5B/* [ */) { return false; }
 
   labelStart = state.pos + 1;
   labelEnd = state.md.helpers.parseLinkLabel(state, state.pos, true);