--- a/lib/rules_inline/newline.js
+++ b/lib/rules_inline/newline.js
@@ -6,11 +6,13 @@ var isSpace = require('../common/utils').isSpace;
 
 
 module.exports = function newline(state, silent) {
-  var pmax, max, pos = state.pos;
+  if (state.src.charCodeAt(state.pos) !== 0x0A/* \n */) { return false; }
 
-  if (state.src.charCodeAt(pos) !== 0x0A/* \n */) { return false; }
+  var pmax, max,
+      pending = state.pending,
+      pos = state.pos;
 
-  pmax = state.pending.length - 1;
+  pmax = pending.length - 1;
   max = state.posMax;
 
   // '  \n' -> hardbreak
@@ -18,12 +20,12 @@ module.exports = function newline(state, silent) {
   // Pending string is stored in concat mode, indexed lookups will cause
   // convertion to flat mode.
   if (!silent) {
-    if (pmax >= 0 && state.pending.charCodeAt(pmax) === 0x20) {
-      if (pmax >= 1 && state.pending.charCodeAt(pmax - 1) === 0x20) {
-        state.pending = state.pending.replace(/ +$/, '');
+    if (pmax >= 0 && pending.charCodeAt(pmax) === 0x20) {
+      if (pmax >= 1 && pending.charCodeAt(pmax - 1) === 0x20) {
+        state.pending = pending.replace(/ +$/, '');
         state.push('hardbreak', 'br', 0);
       } else {
-        state.pending = state.pending.slice(0, -1);
+        state.pending = pending.slice(0, -1);
         state.push('softbreak', 'br', 0);
       }