--- a/lib/rules_inline/strikethrough.js
+++ b/lib/rules_inline/strikethrough.js
@@ -6,13 +6,13 @@
 // Insert each marker as a separate text token, and add it to delimiter list
 //
 module.exports.tokenize = function strikethrough(state, silent) {
-  var i, scanned, token, len, ch,
-      start = state.pos,
-      marker = state.src.charCodeAt(start);
-
   if (silent) { return false; }
 
+  var marker = state.src.charCodeAt(state.pos);
+
   if (marker !== 0x7E/* ~ */) { return false; }
+
+  var i, scanned, token, len, ch;
 
   scanned = state.scanDelims(state.pos, true);
   len = scanned.length;