--- a/lib/rules_inline/text_collapse.js
+++ b/lib/rules_inline/text_collapse.js
@@ -10,26 +10,27 @@
 
 
 module.exports = function text_collapse(state) {
-  var curr, last,
+  var curr, last, token,
       level = 0,
       tokens = state.tokens,
-      max = state.tokens.length;
+      max = tokens.length;
 
   for (curr = last = 0; curr < max; curr++) {
+    token = tokens[curr];
     // re-calculate levels after emphasis/strikethrough turns some text nodes
     // into opening/closing tags
-    if (tokens[curr].nesting < 0) level--; // closing tag
-    tokens[curr].level = level;
-    if (tokens[curr].nesting > 0) level++; // opening tag
+    if (token.nesting < 0) level--; // closing tag
+    token.level = level;
+    if (token.nesting > 0) level++; // opening tag
 
-    if (tokens[curr].type === 'text' &&
+    if (token.type === 'text' &&
         curr + 1 < max &&
         tokens[curr + 1].type === 'text') {
 
       // collapse two adjacent text nodes
-      tokens[curr + 1].content = tokens[curr].content + tokens[curr + 1].content;
+      tokens[curr + 1].content = token.content + tokens[curr + 1].content;
     } else {
-      if (curr !== last) { tokens[last] = tokens[curr]; }
+      if (curr !== last) { tokens[last] = token; }
 
       last++;
     }