I tried to optimize the library a bit, there wasn't much that could have been easily optimized away but there was something.

I don't expect this PR to get merged in its entirety, or at all, I'm mostly writing this as a brain dump, but there are actually a few very minor changes I found that make some signficant difference, maybe those should be ported over here, your choice.

##### Summary

Overall the profile script now runs in ~23% less time.

All tests conducted under Node v15.5.0, older versions might give considerably different results.

<details>
  <summary>10 sample runs (Before)</summary>

```
profile: 2.516s
profile: 2.537s
profile: 2.695s
profile: 2.658s
profile: 2.508s
profile: 2.537s
profile: 2.572s
profile: 2.526s
profile: 2.536s
profile: 2.587s

avg: 2.567s
improvement: 0%
```

</details>

<details>
  <summary>10 sample runs (After)</summary>

```
profile: 1.965s
profile: 1.940s
profile: 1.924s
profile: 1.920s
profile: 2.085s
profile: 2.063s
profile: 1.929s
profile: 1.948s
profile: 1.935s
profile: 1.979s

avg: 1.969s
improvement: 23.3%
```

</details>

##### Optimizations

- Significant:
  - `lib/rules_core/normalize.js`
    - optimizing this single rule gave by far the largest performance gains of the lot. 
    - turns out that replacing "\n" with "\n" is a significant performance burden, so just removing "|\n" from the regex, a 3 characters diff, would be a significant performance improvement.
    - because of a [bug](https://bugs.chromium.org/p/v8/issues/detail?id=10737&sort=-stars&q=&can=4) in v8 looking for "\0" in strings that don't contain only ASCII character is a significant performance cliff, just iterating over the entire thing and performing the change manually turned out to be significantly faster.
      - by the way I don't see the point of replacing "\0" in JS so this replacement should probably be removed entirely, breaking away slightly from CM as this requirement by the spec has only downsides in JS really. 
- Minor:
  - `lib/common/utils.js`
    - Using `String.prototype.repeat` whenever possible, that's 10x faster than the previous implementation and more memory efficient.
  - `lib/renderer.js`
    - `fence`
      - most of the times the "info" string doesn't contain any spaces, so I added a fast path for that that doesn't do so much wasteful stuff.
      - asking for the index of `<pre` is not actually what we need, we don't want to scan the entire thing, we only care about how it starts, so it now only considers the first 4 characters of the string.
      - when adding the "class" attribute pretty much always there's no need to clone the current attributes array at all, so I added a more optimized path for that common case.
  - `lib/rules_block/state_block.js`
    - apparently creating an intermediary array is slower than just progressively joining those lines (~90ms vs ~15ms).
- Very Minor:
  - some unnecessary property accesses have been removed.
  - some exit conditions in rules have been moved slightly up in the functions, ensuring those functions exit as quickly as possible when those exit conditions are met.
    - this was a pretty noisy change with not that much to gain from really.

##### Other Changes

- `benchmark/benchloop.js`
  - I added another profiling script, one that's not so tied to how the CM spec is written and that gives some output quickly.
- `benchmark/profile.js`
  - 10x more iterations are executed, giving more time to v8 to optimize things.
  - A number is outputted, otherwise this file is not very useful IMO, timing the entire script is not what we want as that would count disk reads etc.

##### Potential Future Optimizations

- Near future
  - something like 30% of the time spent is now spent html-escaping strings, maybe that could be optimized somehow, everything I tried so far didn't work. Maybe the set of escaped characters could be shrinked and that would improve things.
  - some functions are not monomorphic, like numbers are being sent to be html-escaped too. It's difficult to optimize on this without any types at all, perhaps if there's any plan to add types to markdown-it there's something that could be done here.
  - the shape of tokens ("hidden class" as v8 calls it) changes too at runtime, I'm not sure how much there is to gain from never changing it, but it may be worth trying something in that regard.
  - overall I just don't think that there are some non-high-effort optimizations that would make the library twice as fast from what it is now.
- Far future
  - at the end of the day dealing with strings seems to be the bottleneck here, we just don't have enough control over them to make manipulating them faster. Perhaps the entire Markdown string should be transformed to bytes, those bytes/numbers should be manipulated directly, and then they could be transformed back to strings at the end of the process.
  - if the compiler worked mostly just with numbers, and didn't use regexes, perhaps it could be ported to WASM with some reasonable amount of effort.
  - if the compiler worked from WASM there might be something interesting that would be possible to do there with SIMD instructions, there seem to be at least a Markdown [compiler](https://github.com/raphlinus/pulldown-cmark) that uses them.

##### End

I'd probably just port over the modified "normalize" rule to get most of the performance benefits with ~none of the noise.

Sorry for the long write.