@puzrin 2021-01-04
> I'd probably just port over the modified "normalize" rule to get most of the performance benefits with ~none of the noise.

Sounds reasonable.

- `\n\r` regexp worth to impove.
- `\0` - not sure. May be, `if \0 exists => call .replace`. Curent implementation cause sliced string store - potential roblem with speed for next scans.