@fabiospampinato 2021-01-04
> \0 - not sure. May be, if \0 exists => call .replace. Curent implementation cause sliced string store - potential roblem with speed for next scans.

I think pretty much always `\0` won't be found in the string anyway, so further optimizations on this might not be that useful, plus `\0` in strings are a non-issue in JS so I'd remove the replacement entirely and squeeze some more performance out of the library. 

Removing the replacement entirely would only cause issues to people reading markdown-it's output from languages like C, and only if the input string contained a `\0` to being with basically, I don't think that happens in practice and those folks have always the option to perform the replacement themselves.

Just my opinion on this, maybe being fully compliant even with the weird parts of CM is worth more.