@puzrin 2021-01-04
If \0 replace is from spec requirement, it can not be dropped. I don't remember details.