@fabiospampinato 2021-01-04
Yeah the spec mentions somewhere that it must be replaced for "security reasons" IIRC, perhaps you could nudge the CM people to change the wording on that in v0.30 of the spec. There are really no security issues with it whatsoever in JS 🤷‍♂️