@j2L4e 2021-01-12
> There are really no security issues with it whatsoever in JS

It depends on what you do with the output and how you display or store it, though. If `\0` sensitive services are involved, this might very well be an issue.

I think an option `replaceNullChars` with default `true` would do. It would also be a non-breaking change and could be documented properly (being not spec-compliant and potentially dangerous)