@fabiospampinato 2021-01-12
> It depends on what you do with the output and how you display or store it, though. If \0 sensitive services are involved, this might very well be an issue.

By that logic if you think about it any string could be a security issue.

The reality is that this is a JS library, `\0`s aren't a problem at all in JS strings, and if other languages have some problems with those they should be the ones performing the sanitization on untrusted things. This isn't a C string sanitization library.

> I think an option replaceNullChars with default true would do. It would also be a non-breaking change and could be documented properly (being not spec-compliant and potentially dangerous)

That'd be too obscure and ugly IMO, plus there kind of is an option for this already since the "normalize" rule can be overridden entirely, but I would imagine approximately 0 people would even know that they can get a double-digit performance percentage boost by not replacing null characters.