@hukkinj1 2021-01-13
> If \0 sensitive services are involved, this might very well be an issue.

> > By that logic if you think about it any string could be a security issue.

I think @j2L4e main argument wasn't really that there is a security issue. Just that some services may be sensitive to changes that are not CommonMark compliant. I myself have written code that relies on the assumption that Markdown parsed by markdown-it can never contain a null character, and I don't think I'm the only one.

Markdown-it advertises 100% CommonMark support in big fat letters so IMO any performance optimizations should be subordinate.