@fabiospampinato 2021-01-26
> I myself have written code that relies on the assumption that Markdown parsed by markdown-it can never contain a null character, and I don't think I'm the only one.

Then you are doing it wrong IMO, for multiple reasons:

- CommonMark can change, if they word the silly requirement about the `\0` character slightly differently then markdown-it may stop replacing that, breaking whatever you are doing with it.
- Markdown-it is not infallible, every now and then there are bugs, how do you even know that the `\0` character is being replaced correctly every single time? Approximately 0 people would dig through the codebase to make sure of that.
- If you are passing the string compiled by markdown-it to a service that is sensitive to null characters then you should be treating that string as an unsafe input in the first place and perform the sanitization right before passing on the string, not doing that leaves you open to potential issues, a couple of which I mentioned in the other bullet points.