@hukkinj1 2021-01-26
> Then you are doing it wrong IMO

If you say so, lol :smile: 

> CommonMark can change

Yup, it's very normal that in case the spec changes, then code does as well, to follow the updated spec. Should I build code that is prepared for any possible spec change? I'm not a clairvoyant, don't think that's possible.

> Markdown-it is not infallible, every now and then there are bugs

So I should build my code with the assumption that all downstream libraries are buggy. And not only that, but predict what bugs exactly will be introduced, and write code that is prepared beforehand?

> f you are passing the string compiled by markdown-it to a service that is sensitive to null characters...

I'm not.