@fabiospampinato 2021-01-26
> Should I build code that is prepared for any possible spec change? I'm not a clairvoyant, don't think that's possible.

I'm only referring to the `\0` issue here, so yes, absolutely, if that's a problem for your use case you should be handling that outside of markdown-it's scope, regardless of what markdown-it does.

> So I should build my code with the assumption that all downstream libraries are buggy. And not only that, but predict what bugs exactly will be introduced, and write code that is prepared beforehand?

Again, regarding the `\0` issue, which is the only thing I was proposing markdown-it broke away from CM for, yes, you should be doing the replacement yourself.

> > f you are passing the string compiled by markdown-it to a service that is sensitive to null characters...
> I'm not.

Well I guess I misunderstood this sentence then:

> I myself have written code that relies on the assumption that Markdown parsed by markdown-it can never contain a null character, and I don't think I'm the only one.

It's good that what markdown-it does with `\0` characters is not a problem anymore for you, that's all I was suggesting, not relying on markdown-it for something like that.