@hukkinj1 2021-01-26
> I'm only referring to the \0 issue here, so yes, absolutely, if that's a problem for your use case you should be handling that outside of markdown-it's scope, regardless of what markdown-it does.

So when markdown-it declares itself 100% CommonMark compliant, I should still make my own calls on what part of the spec its actually not? :smile: 

I'm not saying markdown-it can't stop removing null characters. I'm saying if it does, that should be a breaking major version, and it should be advertised 99% CommonMark compliant afterwards.