@fabiospampinato 2021-01-26
> So when markdown-it declares itself 100% CommonMark compliant, I should still make my own calls on what part of the spec its actually not? 😄

When concerning security issues like the sanitization of null characters, yes, of course.

> I'm not saying markdown-it can't stop removing null characters. I'm saying if it does, that should be a breaking major version, and it should be advertised 99% CommonMark compliant afterwards.

It seems the only way markdown-it could make that change would be if CommonMark itself changed in the first place, in which case it will still be ~100% compliant.